+++ pallets/broadcast/src/tests/mod.rs
@@ -1,2 +1,2 @@
-mod incremental_id;
+mod tests;
 pub mod mock;
